refactor(transaction): move hooks to assets folder
