+++ src/modules/transactions/hooks/index.ts
@@ -1,5 +1,4 @@
+export * from './assets';
 export * from './create';
 export * from './details';
 export * from './list';
-export * from './useFormatSummaryAssets';
-export * from './useGetAssetsByOperations';
